Add journal quantile chart specification (needs some work on the X axis)
